redux dev tools, migrate permanent delete directive
