+++ src/app/components/trashcan/trashcan.component.html
@@ -8,7 +8,6 @@
                 color="primary"
                 mat-icon-button
                 [acaPermanentDelete]="selectedNodes"
-                (selection-node-deleted)="reload()"
                 title="{{ 'APP.ACTIONS.DELETE_PERMANENT' | translate }}">
                 <mat-icon>delete_forever</mat-icon>
             </button>
