+++ src/app/store/actions/node.action.ts
@@ -3,6 +3,12 @@ import { Action } from '@ngrx/store';
 export const SET_SELECTED_NODES = 'SET_SELECTED_NODES';
 export const DELETE_NODES = 'DELETE_NODES';
 export const RESTORE_DELETED_NODES = 'RESTORE_DELETED_NODES';
+export const PURGE_DELETED_NODES = 'PURGE_DELETED_NODES';
+
+export interface NodeInfo {
+    id: string;
+    name: string;
+}
 
 export class SetSelectedNodesAction implements Action {
     readonly type = SET_SELECTED_NODES;
@@ -17,4 +23,9 @@ export class DeleteNodesAction implements Action {
 export class RestoreDeletedNodesAction implements Action {
     readonly type = RESTORE_DELETED_NODES;
     constructor(public payload: any[] = []) {}
+}
+
+export class PurgeDeletedNodesAction implements Action {
+    readonly type = PURGE_DELETED_NODES;
+    constructor(public payload: NodeInfo[] = []) {}
 }
